services/fingerprint/requirements.txt
fastapi
gunicorn
uvicorn[standard]
google-cloud-aiplatform
google-cloud-storage
google-cloud-bigquery
google-cloud-pubsub
opencv-python
Pillow
numpy
requests
pytest
